doing request accept

+++ src/main/java/by/epam/dmitriytomashevich/javatr/courses/command/user/RemoveRequestCommand.java
@@ -8,7 +8,11 @@ import by.epam.dmitriytomashevich.javatr.courses.domain.User;
 import by.epam.dmitriytomashevich.javatr.courses.exceptions.LogicException;
 import by.epam.dmitriytomashevich.javatr.courses.factory.ServiceFactory;
 import by.epam.dmitriytomashevich.javatr.courses.logic.RequestService;
+import com.fasterxml.jackson.databind.node.JsonNodeFactory;
+import com.fasterxml.jackson.databind.node.ObjectNode;
 
+import java.io.IOException;
+import java.io.PrintWriter;
 import java.util.Optional;
 
 public class RemoveRequestCommand implements Command {
@@ -21,11 +25,29 @@ public class RemoveRequestCommand implements Command {
     @Override
     public Optional<String> execute(SessionRequestContent content) throws LogicException {
         User user = (User) content.getSession(false).getAttribute(Parameter.USER);
-        Long conferenceId = Long.valueOf(content.getParameter("conferenceId"));
-        requestService.deleteFullRequestByConferenceIdAndUserId(conferenceId, user.getId());
+        String conferenceIdAsString = content.getParameter("conferenceId");
+        String message = null;
+        boolean isPositiveResult = false;
+        if(conferenceIdAsString!=null){
+            Long conferenceId = Long.valueOf(content.getParameter("conferenceId"));
+            requestService.deleteFullRequestByConferenceIdAndUserId(conferenceId, user.getId());
+            message = "Your request was successfully removed!";
+            isPositiveResult=true;
+        }else {
+            message = "Sorry, something was wrong!";
+        }
 
-        String removeRequestMessage = "Your request was successfully removed!";
-        return Optional.of(ActionNames.PROFILE_ACTION + "?removeRequestMessage=" + removeRequestMessage);
-
+        try {
+            content.getResponse().setContentType("application/json;charset=UTF-8");
+            final JsonNodeFactory factory = JsonNodeFactory.instance;
+            final ObjectNode node = factory.objectNode();
+            node.put("message", message);
+            node.put("isPositiveResult", isPositiveResult);
+            PrintWriter writer = content.getResponse().getWriter();
+            writer.print(node);
+        } catch (IOException e) {
+            throw new LogicException(e);
+        }
+        return Optional.empty();
     }
 }
